+++ Example/HCaptcha/ViewController.swift
@@ -30,6 +30,7 @@ class ViewController: UIViewController {
     @IBOutlet private weak var label: UILabel!
     @IBOutlet private weak var spinner: UIActivityIndicatorView!
     @IBOutlet private weak var localeSegmentedControl: UISegmentedControl!
+    @IBOutlet private weak var apiSegmentedControl: UISegmentedControl!
     @IBOutlet private weak var visibleChallengeSwitch: UISwitch!
     @IBOutlet private weak var validateButton: UIButton!
     @IBOutlet private weak var resetButton: UIButton!
@@ -37,18 +38,6 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         setupHCaptcha()
-
-        if unitTesting {
-            validateButton.isEnabled = false
-            localeSegmentedControl.isEnabled = false
-            visibleChallengeSwitch.isEnabled = false
-        }
-
-        resetButton.rx.tap
-            .subscribe(onNext: { [weak hcaptcha] _ in
-                hcaptcha?.reset()
-            })
-            .disposed(by: disposeBag)
     }
 
     @IBAction func didPressLocaleSegmentedControl(_ sender: UISegmentedControl) {
@@ -62,17 +51,54 @@ class ViewController: UIViewController {
         setupHCaptcha()
     }
 
-    // swiftlint:disable function_body_length
-    @IBAction private func didPressButton(button: UIButton) {
-        if unitTesting {
-            return
+    @IBAction private func didPressVerifyButton(button: UIButton) {
+        if self.apiSegmentedControl.selectedSegmentIndex == 0 {
+            self.verifyRxApi()
+        } else {
+            self.verifyRegularApi()
         }
+    }
 
+    private func verifyRegularApi() {
+        hcaptcha.validate(on: self.view) { result in
+            do {
+                self.label.text = try result.dematerialize()
+            } catch let error as HCaptchaError {
+                self.label.text = error.description
+            } catch let error {
+                self.label.text = String(describing: error)
+            }
+            let subview = self.view.viewWithTag(Constants.webViewTag)
+            subview?.removeFromSuperview()
+        }
+    }
+
+    // swiftlint:disable function_body_length
+    private func verifyRxApi() {
         disposeBag = DisposeBag()
 
         hcaptcha.rx.didFinishLoading
             .debug("did finish loading")
             .subscribe()
+            .disposed(by: disposeBag)
+
+        hcaptcha.rx.events()
+            .debug("events")
+            .subscribe { [weak self] (event, data) in
+                if event == .error {
+                    let error = data as? HCaptchaError
+                    print("onEvent error: \(String(describing: error))")
+                }
+
+                let alertController = UIAlertController(title: "On Event",
+                                                        message: event.rawValue,
+                                                        preferredStyle: .alert)
+                self?.present(alertController, animated: true) {
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+                        alertController.dismiss(animated: true, completion: nil)
+                    }
+                }
+            }
             .disposed(by: disposeBag)
 
         let validate = hcaptcha.rx.validate(on: view, resetOnError: false)
@@ -97,7 +123,7 @@ class ViewController: UIViewController {
             .share(replay: 1)
 
         isEnabled
-            .bind(to: button.rx.isEnabled)
+            .bind(to: validateButton.rx.isEnabled)
             .disposed(by: disposeBag)
 
         validate
@@ -124,6 +150,14 @@ class ViewController: UIViewController {
                 hcaptcha?.reset()
             })
             .disposed(by: disposeBag)
+    }
+
+    @IBAction private func didPressStopButton(button: UIButton) {
+        hcaptcha.stop()
+    }
+
+    @IBAction private func didPressResetButton(button: UIButton) {
+        hcaptcha?.reset()
     }
 
     private func setupHCaptcha() {