+++ Example/ObjC/ViewController.m
@@ -2,8 +2,7 @@
 //  ViewController.m
 //  HCaptcha_Objc_Example
 //
-//  Created by CAMOBAP on 12/27/21.
-//  Copyright © 2021 HCaptcha. MIT License.
+//  Copyright © 2022 HCaptcha. MIT License.
 //
 
 #import "ViewController.h"
@@ -30,6 +29,15 @@
     [self.hCaptcha configureWebView:^(WKWebView * _Nonnull webView) {
         webView.frame = self.view.bounds;
         self.webView = webView;
+    }];
+
+    [self.hCaptcha onEvent:^(enum HCaptchaEvent event, id _Nullable _) {
+        UIAlertController *alert = [UIAlertController alertControllerWithTitle:@"On Event" message: [NSString stringWithFormat:@"%ld", event, nil] preferredStyle:UIAlertControllerStyleAlert];
+        [self presentViewController:alert animated: YES completion:^{
+            dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
+                [alert dismissViewControllerAnimated:YES completion:nil];
+            });
+        }];
     }];
 }